# backend/openapi/openapi.yml
openapi: 3.0.3
info:
  title: "UECS Navi API"
  version: "1.0.0"
tags:
  - name: Utilities
  - name: DeviceSettingPage
  - name: CreateWorkflowPage
  - name: TimeSchedulePage
paths:
  /climate-datas:
    get:
      operationId: getClimateData
      tags:
        - Utilities
      summary: Get climate data list
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClimateData"

  /m304s:
    get:
      operationId: getM304s
      tags:
        - Utilities
      summary: Get M304 ID list
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/M304ID"

  /m304s/{m304-id}/devices:
    parameters:
      - $ref: "#/components/parameters/m304-id"
    get:
      operationId: getDevices
      tags:
        - Utilities
      summary: Get devices connected to M304
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceResponse"

  /houses:
    get:
      operationId: getHouses
      tags:
        - DeviceSettingPage
      summary: Get houses list
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HousesResponse"
                  
  /houses/{house-id}/devices:
    parameters:
      - $ref: "#/components/parameters/house-id"
    post:
      operationId: createDevice
      tags:
        - DeviceSettingPage
      summary: Create device
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceRequest"
      responses:
        "200":
          description: OK

  /workflows-with-ui/{m304-id}:
    parameters:
    - $ref: "#/components/parameters/m304-id"
    get:
      operationId: getWorkflowsWithUI
      tags:
        - CreateWorkflowPage
      summary: Get workflows with the state of React Flow.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowWithUIResponse"

  /workflows-with-ui:
    post:
      operationId: createWorkflowWithUI
      tags:
        - CreateWorkflowPage
      summary: Create a workflow with UI
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowWithUIRequest"
      responses:
        "200":
          description: OK
    put:
      operationId: updateWorkflowWithUI
      tags:
        - CreateWorkflowPage
      summary: Update a workflow with UI
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowWithUIRequest"
      responses:
        "200":
          description: OK
  /workflow-with-ui/{workflow-id}:
    parameters:
      - $ref: "#/components/parameters/workflow-id"
    delete:
      operationId: deleteWorkflowWithUI
      tags:
        - CreateWorkflowPage
      summary: Delete a workflow with UI
      responses:
        "200":
          description: OK

  /time-schedule/{m304-id}:
    parameters:
      - $ref: "#/components/parameters/m304-id"
    get:
      operationId: getTimeSchedule
      tags:
        - TimeSchedulePage
      summary: Get time schedule
      requestBody:
          content:
            application/json:
              schema:
                 $ref: "#/components/schemas/M304IDRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeScheduleResponse"

  /time-schedule:
    post:
      operationId: createAndBuildTimeSchedule
      tags:
      - TimeSchedulePage
      summary: Create and build time schedule
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeScheduleRequest"
      responses:
        "200":
          description: OK
    put: 
      operationId: updateAndBuildTimeSchedule
      tags:
        - TimeSchedulePage
      summary: Update and build time schedule
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeScheduleRequest"
      responses:
        "200": 
          description: OK

components:
  parameters:
    house-id:
      name: house-id
      in: path
      required: true
      schema:
        type: integer
    m304-id:
      name: m304-id
      in: path
      required: true
      schema:
        type: integer
    workflow-id:
      name: workflow-id
      in: path
      required: true
      schema:
        type: integer

  schemas:
    # General schemas
    ClimateData:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        unit:
          type: string

    WorkflowOperations:
      type: object
      properties:
        relay_1:
          type: boolean
        relay_2: 
          type: boolean
        relay_3: 
          type: boolean
        relay_4: 
          type: boolean
        relay_5: 
          type: boolean
        relay_6: 
          type: boolean
        relay_7: 
          type: boolean
        relay_8: 
          type: boolean
    
    Workflow:
      type: object
      properties:
        id:
          type: integer
        m304_id:
          type: integer
        name:
          type: string
    
    Condition:
      type: object
      properties:
        selected_climate_data_id:
          type: integer
        selected_comparison_operator_id:
          type: integer
        set_point:
          type: number
          format: float
    
    TimeScheduleRow:
      type: object
      properties:
        start_time:
          type: string
          pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
          description: "開始時刻（HH:mm形式）"
        end_time:
          type: string
          pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
          description: "終了時刻（HH:mm形式）"
        selected_workflow_id:
          type: integer
        condition:
            $ref: "#/components/schemas/Condition"

    M304ID: 
      type: integer

    # Requests
    HousesRequest:
      type: string
    
    M304IDRequest:
      type: integer

    DeviceRequest:
      type: object
      properties:
        climate_data_id:
          type: integer
        m304_id:
          type: integer
        name:
          type: string
        rly:
          type: integer

    NodeRequest: 
      type: object
      properties:
        workflow_node_id:
          type: string
        type: 
          type: string
        data: 
          type: object
          additionalProperties: true
        position_x:
          type: number
          format: float
        position_y:
          type: number
          format: float

    EdgeRequest:
      type: object
      properties:
        source_node_id:
          type: string
        target_node_id:
          type: string

    WorkflowUIRequest:
          type: object
          properties:
            nodes: 
              type: array
              items:
                $ref: "#/components/schemas/NodeRequest"
            edges:
              type: array
              items:
                $ref: "#/components/schemas/EdgeRequest"

    WorkflowWithUIRequest:
      type: object
      properties:
        workflow:
          $ref: "#/components/schemas/Workflow"
        workflowUI:
          $ref: "#/components/schemas/WorkflowUIRequest"
        relays:
          $ref:
            "#/components/schemas/WorkflowOperations"

    TimeScheduleRequest:
      type: object
      properties:
        m304_id:
          type: integer
        time_schedule:
          type: array
          items: 
            $ref: "#/components/schemas/TimeScheduleRow"

    # Response schemas
    HousesResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string

    DeviceResponse:
      type: object
      properties:
        id:
          type: integer
        climate_data_id:
          type: integer
        m304_id:
          type: integer
        name:
          type: string
        rly:
          type: integer

    NodeResponse: 
      type: object
      properties:
        id:
          type: integer
        workflow_id:
          type: integer
        workflow_node_id:
          type: string
        type: 
          type: string
        data: 
          type: object
          additionalProperties: true
        position_x:
          type: number
          format: float
        position_y:
          type: number
          format: float

    EdgeResponse:
      type: object
      properties:
        id:
          type: integer
        workflow_id:
          type: integer
        source_node_id:
          type: string
        target_node_id:
          type: string
          
    WorkflowUIResponse:
          type: object
          properties:
            nodes: 
              type: array
              items:
                $ref: "#/components/schemas/NodeResponse"
            edges:
              type: array
              items:
                $ref: "#/components/schemas/EdgeResponse"

    WorkflowWithUIResponse:
      type: object
      properties:
        workflow:
          $ref: "#/components/schemas/Workflow"
        workflowUI:
          $ref: "#/components/schemas/WorkflowUIResponse"

    TimeScheduleResponse:
      type: object
      properties:
        id: 
          type: integer
        m304_id:
          type: integer
        workflows:
          type: array
          items:
            $ref: "#/components/schemas/Workflow"
        time_schedules:
          type: array
          items:
            $ref: "#/components/schemas/TimeScheduleRow"